2026/02/06 15:00 강우준

+++ src/main/java/com/drawing/springboot/dto/SubcategoryDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class SubcategoryDTO {
-	private Long subcategoryId;
-    private Long categoryId;
+	private int subcategoryId;
+    private int categoryId;
     private String name;
     private String url;
     private String image;
